perf: reduce cold imports during provider discovery
