+++ src/agents/models-config.providers.cloudflare-ai-gateway.test.ts
@@ -1,8 +1,8 @@
 // Covers Cloudflare AI Gateway profile provenance and generated provider config.
 import { describe, expect, it } from "vitest";
+import { NON_ENV_SECRETREF_MARKER } from "../secrets/provider-credential-values.js";
 import { captureEnv } from "../test-utils/env.js";
 import type { ApiKeyCredential } from "./auth-profiles/types.js";
-import { NON_ENV_SECRETREF_MARKER } from "./model-auth-markers.js";
 import { resolveApiKeyFromCredential } from "./models-config.providers.secret-helpers.js";
 
 function expectedCloudflareGatewayBaseUrl(accountId: string, gatewayId: string): string {
