+++ src/agents/models-config.providers.source-managed.ts
@@ -4,9 +4,9 @@ import { normalizeProviderId } from "@openclaw/model-catalog-core/provider-id";
  */
 import { resolveConfigSecretRef } from "../config/resolution-facts.js";
 import type { OpenClawConfig } from "../config/types.openclaw.js";
+import { resolveNonEnvSecretRefApiKeyMarker } from "../secrets/provider-credential-values.js";
 import { isRecord } from "../utils.js";
 import {
-  resolveNonEnvSecretRefApiKeyMarker,
   resolveNonEnvSecretRefHeaderValueMarker,
   resolveEnvSecretRefHeaderValueMarker,
 } from "./model-auth-markers.js";
