+++ src/agents/tools/web-search-provider-common.ts
@@ -73,16 +73,6 @@ export function readConfiguredSecretString(value: unknown, path: string): string
   return normalizeSecretInput(normalizeResolvedSecretInputString({ value, path })) || undefined;
 }
 
-export function readProviderEnvValue(envVars: string[]): string | undefined {
-  for (const envVar of envVars) {
-    const value = normalizeSecretInput(process.env[envVar]);
-    if (value) {
-      return value;
-    }
-  }
-  return undefined;
-}
-
 export async function withTrustedWebSearchEndpoint<T>(
   params: WebSearchEndpointOptions,
   run: (response: Response) => Promise<T>,
