+++ src/commands/models/list.auth-overview.secret-ref.test.ts
@@ -1,10 +1,10 @@
 import { afterEach, describe, expect, it } from "vitest";
-import { NON_ENV_SECRETREF_MARKER } from "../../agents/model-auth-markers.js";
 import {
   clearRuntimeConfigSnapshot,
   setRuntimeConfigSnapshot,
 } from "../../config/runtime-snapshot.js";
 import type { OpenClawConfig } from "../../config/types.js";
+import { NON_ENV_SECRETREF_MARKER } from "../../secrets/provider-credential-values.js";
 import { resolveProviderAuthOverview } from "./list.auth-overview.js";
 
 const credential = "synthetic-resolved-provider-credential";
