+++ src/plugin-sdk/provider-auth-runtime.ts
@@ -18,7 +18,7 @@ export {
   collectProviderApiKeysForExecution,
   executeWithApiKeyRotation,
 } from "../agents/api-key-rotation.js";
-export { NON_ENV_SECRETREF_MARKER } from "../agents/model-auth-markers.js";
+export { NON_ENV_SECRETREF_MARKER } from "../secrets/provider-credential-values.js";
 export {
   isProviderAuthError,
   requireApiKey,
